[encoder] rename mb->partition; impl block and partition, sb alloc and dealloc; reset enc structs; [tx] add size lookup; [mem] assert malloc; bigger sz; [pix] impl dup view, from channel; [test] add test block data struct alloc and dealloc

+++ src/util/mem.h
@@ -11,7 +11,7 @@ extern "C" {
 
 // memory pool
 struct c1_mpool_s {
-    const uint8_t sz; // elem_sz
+    const uint16_t sz; // elem_sz
     const uint8_t nb;
     void *root_;
 };
